app/requirements.txt
# --- base dependecies ---
# flask
# flask-sqlalchemy
# gunicorn
# redis
# celery
# pycryptodome

amqp==2.5.2
billiard==3.6.2.0
celery==4.4.0
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.7
MarkupSafe==1.1.1
pycryptodome==3.9.6
pytz==2019.3
redis==3.4.1
SQLAlchemy==1.3.13
vine==1.3.0
Werkzeug==1.0.0
